refactor(ui): extract time entry form to component
